CCBUGS: 266175 Some more on average calculation (more usage of float instead of long).
svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1254102

+++ core/mixdevicecomposite.cpp
@@ -83,7 +83,7 @@ long MixDeviceComposite::calculateVolume(Volume::VolumeType vt)
 
         Volume& vol =  ( vt == Volume::CaptureVT ) ? md->captureVolume() : md->playbackVolume();
         if (vol.hasVolume() && (vol.maxVolume() != 0) ) {
-            long normalizedVolume =
+            qreal normalizedVolume =
                       ( vol.getAvgVolume(Volume::MALL) * MixDeviceComposite::VolMax )
                     /   vol.maxVolume();
             volSum += normalizedVolume;
